🎨 Design : CvBasicInfo, CvCarrer UI
