+++ client/src/pages/Root.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from '../components/Header';
-import Footer from '../components/Footer';
+import Header from '../components/Cv/Header';
+import Footer from '../components/Cv/Footer';
 
 export default function Root() {
   return (
